pyproject.toml
[project]
name = "food-language-mistralapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.12.0",
    "ipykernel>=7.1.0",
    "mistralai>=1.9.11",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pip-audit>=2.9.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
]
